Add proper CLI interface & add log output format option
The log output format will by default autodetect if it is currently running inside a Github Action / Workflow and adjust its output format to be readable as an annotation (see https://docs.github.com/en/free-pro-team@latest/actions/reference/workflow-commands-for-github-actions#setting-an-error-message ).

This patch also introduces the -v (verbose) flag to get debugging logs etc. 

And the correct assembly version is now set to 0.1.3.0.